+++ server/crates/arachnea-stream/src/services/mod.rs
@@ -4,3 +4,4 @@ pub(crate) mod player_resolver;
 pub(crate) mod rtbf_auvio_resolver;
 pub(crate) mod rtlplay_resolver;
 pub(crate) mod tf1_resolver;
+pub(crate) mod tv5mondeplus_resolver;